Add Rouns 1085 and 1086, RPCs 2026 01 and 02, and reorder RPCs

+++ CodeForces/Contest/2025/April Fool Days 2025/J.cpp
@@ -7,7 +7,7 @@ using ld = long double;
 #define sz size()
 #define fi first
 #define se second
-typedef pair<int, int> pii;
+typedef pair<ll, ll> pii;
 
 void solver(){
     string s; cin>>s;
@@ -15,9 +15,9 @@ void solver(){
     else cout<<"Nauuo and Chess"<<endl;
 }
 
-int main(){
+ll main(){
     ios_base::sync_with_stdio(0);cin.tie(NULL);
-    int t = 1;
+    ll t = 1;
     // cin>>t;
     while(t--){
         solver();
