+++ CodeForces/ITMO Academy/DSU/Step 2/J.cpp
@@ -5,11 +5,11 @@ using ll = long long;
 using ld = long double;
 #define pb push_back
 #define sz size()
-typedef pair<int, int> pii;
+typedef pair<ll, ll> pii;
 
-const int maxn = 300005, inf = 2e9;
-int components;
-vector<int> leader(maxn), sizen(maxn), len(maxn);
+const ll maxn = 300005, inf = 2e9;
+ll components;
+vector<ll> leader(maxn), sizen(maxn), len(maxn);
 // En el array len almacenamos la paridad de la distancia de
 // un vertice v a su lider
 
@@ -19,15 +19,15 @@ vector<int> leader(maxn), sizen(maxn), len(maxn);
 // Source: https://codeforces.com/edu/course/2/lesson/7/2
 
 // Implementar DSU modificado para llenar el vector "len"
-void initDSU(int n){
-    for(int i = 1; i <= n; i++){
+void initDSU(ll n){
+    for(ll i = 1; i <= n; i++){
         leader[i] = i;
         sizen[i] = 1;
         len[i] = 0;
     }
 }
 
-pii find(int u){
+pii find(ll u){
     if(leader[u] == u) return {u, 0};
     
     pii val = find(leader[u]);
@@ -37,8 +37,8 @@ pii find(int u){
     // Retorna el lider de u y la paridad de la distancia al lider
 }
 
-int join(int u, int v, int i){
-    int ans = inf;
+ll join(ll u, ll v, ll i){
+    ll ans = inf;
     pii valU = find(u), valV = find(v);
     if(valU.first != valV.first){
         if(sizen[valU.first] > sizen[valV.first])
@@ -55,26 +55,26 @@ int join(int u, int v, int i){
     return ans;
 }
 
-void updateAllLeaders(int n){
-    for(int i = 1; i <= n; i++)
+void updateAllLeaders(ll n){
+    for(ll i = 1; i <= n; i++)
         find(i);
 }
 
 void solver(){
-    int n, m; cin>>n>>m;
+    ll n, m; cin>>n>>m;
     initDSU(n);
-    int ans = inf;
-    for(int i = 1; i <= m; i++){
-        int u, v; cin>>u>>v;
+    ll ans = inf;
+    for(ll i = 1; i <= m; i++){
+        ll u, v; cin>>u>>v;
         ans = min(ans, join(u, v, i));
     }
 
     cout<<(ans == inf ? -1 : ans)<<endl;
 }
 
-int main(){
+ll main(){
     ios_base::sync_with_stdio(0);cin.tie(NULL);
-    int t = 1;
+    ll t = 1;
     // cin>>t;
     while(t--){
         solver();
